Add acrn_common.h
The acrn_mngr.h include acrn_common.h, to get the VM name length,
which is defined by acrn hypervisor now.
And acrn_common.h also included lots of headers in acrn-hypervisor
source code, which make it to be hard to port to acrn_common.h to
another project.
So we can add an acrn_common.h for cbc tools mannully.

Signed-off-by: Tao Yuhong <yuhong.tao@intel.com>